Update README.md
This is the coding for water tank level indicator using Ardunio with C programming. The water level is determined and ultrasonic is fixed with it. A buzzer and led is available along to it. The buzzer will raise a sound when the water level is below 5% and above 25%. The LED indicate that "water is low" when it is below to the range. And the LED indicate that " tank is full" when it is above the range. 

This system will most helpful in apartments where the tank level is don't known and people can determine the level using this device.